Sergey <kubtek@gmail.com>:
SHELL_CMD option in cmake                                                                          
The scripts in the tests directory were written to be proccessed by some kind                         
of bash-compatible shell. On FreeBSD /bin/sh referes to a conventional shell                          
that cannot process these scripts. In particular, the source command is                               
absent. That is why I added an options that allows to specify the shell to                            
use. By default that option is /bin/sh, so everything that worked without this                        
option will work as before. 



git-svn-id: https://xdxf.svn.sourceforge.net/svnroot/xdxf/trunk@97 9cfc4783-1b1d-0410-aaba-dcc21bd7f26c